test(integration-tests/airdrop-pallet): W.I.P.
